Check permission in controller

+++ src/Infrastructure/Http/Rest/Controller/AlbumController.php
@@ -10,6 +10,7 @@ use FOS\RestBundle\Controller\Annotations as Rest;
 use FOS\RestBundle\Controller\FOSRestController;
 use FOS\RestBundle\View\View;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
+use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 
@@ -42,6 +43,15 @@ final class AlbumController extends FOSRestController
      */
     public function postAlbum(AlbumDTO $albumDTO): View
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+        if ($albumDTO->getOwner() == $this->getUser())
+        {
+            $data = [
+                'message' => "Invalid User"
+            ];
+
+            return new JsonResponse($data, Response::HTTP_FORBIDDEN);
+        }
         $album = $this->albumService->addAlbum($albumDTO);
 
         // In case our POST was a success we need to return a 201 HTTP CREATED response with the created object
@@ -87,6 +97,15 @@ final class AlbumController extends FOSRestController
      */
     public function putAlbum(int $albumId, AlbumDTO $albumDTO): View
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+        if ($albumDTO->getOwner() == $this->getUser())
+        {
+            $data = [
+                'message' => "Invalid User"
+            ];
+
+            return new JsonResponse($data, Response::HTTP_FORBIDDEN);
+        }
         $album = $this->albumService->updateAlbum($albumId, $albumDTO);
 
         // In case our PUT was a success we need to return a 200 HTTP OK response with the object as a result of PUT
@@ -102,6 +121,16 @@ final class AlbumController extends FOSRestController
      */
     public function deleteAlbum(int $albumId): View
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+        $album = $this->albumService->getAlbum($albumId);
+        if ($album->get == $this->getUser())
+        {
+            $data = [
+                'message' => "Invalid User"
+            ];
+
+            return new JsonResponse($data, Response::HTTP_FORBIDDEN);
+        }
         $this->albumService->deleteAlbum($albumId);
 
         // In case our DELETE was a success we need to return a 204 HTTP NO CONTENT response. The object is deleted.
